fix bug in seraching for targets
Basically when looking for targets it uses the verb returned from TryGetAttackVerb(null) which I set as the longest ranged verb but if that verb has a minRange then it won't think it can hit anything closer than that. Now, each VerbManager maintains its own SearchVerb that is the widest possible selection of targeting restrictions that it uses for searching for targets.